+++ src/tests/stories/modules/asserts/assertsNull/10a-CanAssertIsNullStory.php
@@ -1,0 +1,31 @@
+<?php
+
+use Storyplayer\SPv2\Modules\Asserts;
+use Storyplayer\SPv2\Modules\Checkpoint;
+use Storyplayer\SPv2\Modules\Log;
+use Storyplayer\SPv2\Stories\BuildStory;
+
+// ========================================================================
+//
+// STORY DETAILS
+//
+// ------------------------------------------------------------------------
+
+$story = BuildStory::newStory();
+
+// ========================================================================
+//
+// ACTIONS
+//
+// ------------------------------------------------------------------------
+
+$story->addAction(function() {
+    // what are we doing?
+    $log = Log::usingLog()->startAction("can assert that data is NULL");
+
+    // this is where you perform the steps of your user story
+    Asserts::assertsNull(null)->isNull();
+
+    // all done
+    $log->endAction();
+});