+++ src/tests/stories/modules/asserts/assertsNull/10b-CannotAssertStringIsNullStory.php
@@ -1,0 +1,47 @@
+<?php
+
+use Storyplayer\SPv2\Modules\Asserts;
+use Storyplayer\SPv2\Modules\Checkpoint;
+use Storyplayer\SPv2\Modules\Failure;
+use Storyplayer\SPv2\Modules\Log;
+use Storyplayer\SPv2\Stories\BuildStory;
+
+// ========================================================================
+//
+// STORY DETAILS
+//
+// ------------------------------------------------------------------------
+
+$story = BuildStory::newStory();
+
+// ========================================================================
+//
+// ACTIONS
+//
+// ------------------------------------------------------------------------
+
+$story->addAction(function() {
+    // what are we doing?
+    $log = Log::usingLog()->startAction("cannot assert empty string is NULL");
+
+    // this is where you perform the steps of your user story
+    Failure::expectsFailure()->when("assertion", function() {
+        Asserts::assertsNull('')->isNull();
+    });
+
+    // all done
+    $log->endAction();
+});
+
+$story->addAction(function() {
+    // what are we doing?
+    $log = Log::usingLog()->startAction("cannot assert 'hello, world' is NULL");
+
+    // this is where you perform the steps of your user story
+    Failure::expectsFailure()->when("assertion", function() {
+        Asserts::assertsNull('hello, world')->isNull();
+    });
+
+    // all done
+    $log->endAction();
+});